feat: warn about unused module members
