+++ packages/safe-ds-lang/src/language/validation/safe-ds-validator.ts
@@ -76,11 +76,7 @@ import {
     constantParameterMustHaveTypeThatCanBeEvaluatedToConstant,
 } from './other/declarations/parameters.js';
 import { placeholderShouldBeUsed, placeholdersMustNotBeAnAlias } from './other/declarations/placeholders.js';
-import {
-    segmentParameterShouldBeUsed,
-    segmentResultMustBeAssignedExactlyOnce,
-    segmentShouldBeUsed,
-} from './other/declarations/segments.js';
+import { segmentParameterShouldBeUsed, segmentResultMustBeAssignedExactlyOnce } from './other/declarations/segments.js';
 import {
     typeParameterMustBeUsedInCorrectPosition,
     typeParameterMustHaveSufficientContext,
@@ -190,6 +186,7 @@ import {
 } from './other/declarations/parameterBounds.js';
 import { unknownMustOnlyBeUsedAsDefaultValueOfStub } from './other/expressions/literals.js';
 import { tagsShouldNotHaveDuplicateEntries } from './builtins/tags.js';
+import { moduleMemberShouldBeUsed } from './other/declarations/moduleMembers.js';
 
 /**
  * Register custom validation checks.
@@ -325,6 +322,7 @@ export const registerValidationChecks = function (services: SafeDsServices) {
             pipelineFileMustNotBeInBuiltinPackage,
             pythonModuleShouldDifferFromSafeDsPackage(services),
         ],
+        SdsModuleMember: [moduleMemberShouldBeUsed(services)],
         SdsNamedType: [
             namedTypeDeclarationShouldNotBeDeprecated(services),
             namedTypeDeclarationShouldNotBeExperimental(services),
@@ -367,7 +365,6 @@ export const registerValidationChecks = function (services: SafeDsServices) {
             segmentParameterShouldBeUsed(services),
             segmentResultMustBeAssignedExactlyOnce(services),
             segmentResultListShouldNotBeEmpty(services),
-            segmentShouldBeUsed(services),
         ],
         SdsStatement: [statementMustDoSomething(services)],
         SdsTemplateString: [templateStringMustHaveExpressionBetweenTwoStringParts],
