+++ packages/safe-ds-lang/tests/resources/validation/other/declarations/module members/unused/same package.sdsdev
@@ -1,4 +1,4 @@
-package tests.validation.other.declarations.segments.unused
+package tests.validation.other.declarations.moduleMembers.unused
 
 pipeline myPipeline2 {
     myUsedInternalSegment();
